# jlink: correctly return WAIT status for AP reads.