Fixes flashing the nrf91 with the onboard jlink. 

Chip erase still fails though. I think it has never worked, even in past versions where flashing did work.

related: #1223