@diondokter 2023-02-03
Ok, let's confirm that this PR is working on my setup.

Current release version:
```
❯ cargo flash --release --chip nrf9160
    Finished release [optimized + debuginfo] target(s) in 0.27s
    Flashing C:\Repos\XXX\target\thumbv8m.main-none-eabihf\release\XXX
        WARN probe_rs::config::registry > Found chip nRF9160_xxAA which matches given partial name nrf9160. Consider specifying its full name.
        WARN probe_rs::config::registry > Matching nrf9160 based on wildcard. Consider specifying the chip as nRF9160_xxAA instead.
        WARN probe_rs::config::target   > Using custom sequence for nRF9160
        WARN probe_rs::session          > Could not clear all hardware breakpoints: Arm(AccessPort { address: ApAddress { dp: Default, ap: 0 }, source: RegisterWrite { address: 0, name: "CSW", source: Dap(FaultResponse) } })
        WARN probe_rs::session          > Could not stop core tracing: Arm(AccessPort { address: ApAddress { dp: Default, ap: 0 }, source: RegisterWrite { address: 0, name: "CSW", source: Dap(FaultResponse) } })                                  
     Erasing sectors ✔ [00:00:00] [----------------------------------------------------------------------------------------------------------------------------------------------------------------------]       0B/316.00 KiB @ 7.10 MiB/s (eta 0s )
 Programming pages   ✔ [00:00:00] [----------------------------------------------------------------------------------------------------------------------------------------------------------------------]       0B/316.00 KiB @ 2.75 MiB/s (eta 0s )       Error Failed to erase flash sector at address 0x00010000.

  Caused by:
          0: Failed to erase flash sector at address 0x00010000.
          1: Failed to erase flash sector at address 0x00010000.
          2: Something during the interaction with the core went wrong
          3: A ARM specific error occured.
          4: Error using access port
          5: Failed to read register DRW at address 0x0000000c
          6: An error occured in the communication with an access port or debug port.
          7: Target device responded with a FAULT response to the request.

        Hint Perhaps your chip has write protected sectors that need to be cleared?

        Hint Perhaps you need the --nmagic linker arg. See https://github.com/rust-embedded/cortex-m-quickstart/pull/95 for more information.
```

With this branch:
```
❯ cargo flash --release --chip nrf9160
    Finished release [optimized + debuginfo] target(s) in 0.28s
    Flashing C:\Repos\XXX\target\thumbv8m.main-none-eabihf\release\XXX
        WARN probe_rs::config::registry > Found chip nRF9160_xxAA which matches given partial name nrf9160. Consider specifying its full name.
        WARN probe_rs::config::registry > Matching nrf9160 based on wildcard. Consider specifying the chip as nRF9160_xxAA instead.
        WARN probe_rs::config::target   > Using custom sequence for nRF9160
     Erasing sectors ✔ [00:00:08] [######################################################################################################################################################################################] 316.00 KiB/316.00 KiB @ 36.82 KiB/s (eta 0s )
 Programming pages   ✔ [00:00:08] [######################################################################################################################################################################################] 316.00 KiB/316.00 KiB @ 38.95 KiB/s (eta 0s )    Finished in 16.768s
```

This makes me incredibly happy!